refactor(docs): drawer dx
