+++ apps/docs/content/components/drawer/custom-motion.ts
@@ -1,58 +1,4 @@
-const App = `import {Drawer, DrawerContent, DrawerHeader, DrawerBody, DrawerFooter, Button, useDisclosure} from "@nextui-org/react";
-
-export default function App() {
-  const {isOpen, onOpen, onOpenChange} = useDisclosure();
-
-  return (
-    <>
-      <Button onPress={onOpen}>Open Drawer</Button>
-      <Drawer 
-        isOpen={isOpen} 
-        onOpenChange={onOpenChange}
-        motionProps={{
-          variants: {
-            enter: {
-              opacity: 1,
-              x: 0,
-              duration: 0.3,
-            },
-            exit: {
-              x: 100,
-              opacity: 0,
-              duration: 0.3,
-            },
-          },
-        }}
-      >
-        <DrawerContent>
-          {(onClose) => (
-            <>
-              <DrawerHeader className="flex flex-col gap-1">Custom Motion Drawer</DrawerHeader>
-              <DrawerBody>
-                <p>
-                  This drawer has custom enter/exit animations.
-                </p>
-                <p> 
-                  Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                  Nullam pulvinar risus non risus hendrerit venenatis.
-                  Pellentesque sit amet hendrerit risus, sed porttitor quam.
-                </p>
-              </DrawerBody>
-              <DrawerFooter>
-                <Button color="danger" variant="light" onPress={onClose}>
-                  Close
-                </Button>
-                <Button color="primary" onPress={onClose}>
-                  Action
-                </Button>
-              </DrawerFooter>
-            </>
-          )}
-        </DrawerContent>
-      </Drawer>
-    </>
-  );
-}`;
+import App from "./custom-motion.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
